% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stream_power.R
\name{stream_power}
\alias{stream_power}
\title{Stream Power}
\usage{
stream_power(Q, S, rho = 1000, g = 9.8)
}
\arguments{
\item{Q}{stream discharge (m^3/s)}

\item{S}{channel slope}

\item{rho}{water density (g/m^3) default = 1000}

\item{g}{acceleration due to gravity (m/s^2) default = 9.8}
}
\value{
stream power (W)
}
\description{
This function determines the power of a stream
}
